repository/requirements.txt
attrs==19.3.0
Automat==20.2.0
constantly==15.1.0
hyperlink==19.0.0
idna==2.10
incremental==17.5.0
numpy==1.19.1
Pillow==6.2.2
pkg-resources==0.0.0
PyHamcrest==2.0.2
six==1.15.0
Twisted==19.10.0
zope.interface==5.1.0
